# xe: sdpa: Optimizations for causal mask. Use split barriers for Q tensor