# Description

This PR optimizes SDPA with causal masks by skipping the processing of masked out values. This includes skipping work-groups and reducing the number of iterations of the inner loop of the micro_sdpa kernel. With this change we are seeing a speedup of up to 1.9x over the same kernel before this change.

Additionally this PR uses split barriers instead of the regular OpenCL barrier when reading Q tenor's values into shared memory. This is only faster for first token cases so I have gated the feature behind a kernel definition.

Fixes: [MFDNN-13683](https://jira.devtools.intel.com/browse/MFDNN-13683)