@petercad 2025-05-19
@umar456 How much benefit did the split barriers provide?