@umar456 2025-05-20
> @umar456 How much benefit did the split barriers provide?

I saw speedups between .89-1.08x over the barrier version of the kernel. There were some notable slowdowns but it seemed to be isolated to second token cases.