@umar456 2025-05-20
Here are the performance improvements based on this change:

BMG
| mb | Q Heads | KV Heads |   D |    K |    Q | Kdt | Vdt | mask     | quant           | base time (us) | PR time (us) |           |
|----|---------|----------|-----|------|------|-----|-----|----------|-----------------|----------------|--------------|-----------|
|  1 |      32 |        8 | 128 |   64 |   64 | f16 | f16 | causaltl | no_quantization |              5 |            4 |      1.25 |
|  1 |      32 |        8 | 128 |   65 |    1 | f16 | f16 | causaltl | no_quantization |              4 |            4 |         1 |
|  1 |      32 |        8 | 128 |  128 |  128 | f16 | f16 | causaltl | no_quantization |             13 |           12 | 1.0833333 |
|  1 |      32 |        8 | 128 |  129 |    1 | f16 | f16 | causaltl | no_quantization |              5 |            5 |         1 |
|  1 |      32 |        8 | 128 |  384 |  384 | f16 | f16 | causaltl | no_quantization |             58 |           45 | 1.2888889 |
|  1 |      32 |        8 | 128 |  385 |    1 | f16 | f16 | causaltl | no_quantization |              9 |            8 |     1.125 |
|  1 |      32 |        8 | 128 |  512 |  512 | f16 | f16 | causaltl | no_quantization |             90 |           61 | 1.4754098 |
|  1 |      32 |        8 | 128 |  513 |    1 | f16 | f16 | causaltl | no_quantization |             12 |           11 | 1.0909091 |
|  1 |      32 |        8 | 128 | 1024 | 1024 | f16 | f16 | causaltl | no_quantization |            314 |          187 | 1.6791444 |
|  1 |      32 |        8 | 128 | 1025 |    1 | f16 | f16 | causaltl | no_quantization |             20 |           18 | 1.1111111 |
|  1 |      32 |        8 | 128 | 2048 | 2048 | f16 | f16 | causaltl | no_quantization |           1228 |          641 | 1.9157566 |
|  1 |      32 |        8 | 128 | 2049 |    1 | f16 | f16 | causaltl | no_quantization |             35 |           33 | 1.0606061 |
|  1 |      32 |        8 | 128 |   64 |   64 | f16 | f16 | causalbr | no_quantization |              5 |            4 |      1.25 |
|  1 |      32 |        8 | 128 |   65 |    1 | f16 | f16 | causalbr | no_quantization |              4 |            4 |         1 |
|  1 |      32 |        8 | 128 |  128 |  128 | f16 | f16 | causalbr | no_quantization |             13 |           12 | 1.0833333 |
|  1 |      32 |        8 | 128 |  129 |    1 | f16 | f16 | causalbr | no_quantization |              5 |            4 |      1.25 |
|  1 |      32 |        8 | 128 |  384 |  384 | f16 | f16 | causalbr | no_quantization |             58 |           45 | 1.2888889 |
|  1 |      32 |        8 | 128 |  385 |    1 | f16 | f16 | causalbr | no_quantization |              9 |            8 |     1.125 |
|  1 |      32 |        8 | 128 |  512 |  512 | f16 | f16 | causalbr | no_quantization |             90 |           61 | 1.4754098 |
|  1 |      32 |        8 | 128 |  513 |    1 | f16 | f16 | causalbr | no_quantization |             12 |           11 | 1.0909091 |
|  1 |      32 |        8 | 128 | 1024 | 1024 | f16 | f16 | causalbr | no_quantization |            312 |          186 | 1.6774194 |
|  1 |      32 |        8 | 128 | 1025 |    1 | f16 | f16 | causalbr | no_quantization |             20 |           19 | 1.0526316 |
|  1 |      32 |        8 | 128 | 2048 | 2048 | f16 | f16 | causalbr | no_quantization |           1227 |          640 | 1.9171875 |
|  1 |      32 |        8 | 128 | 2049 |    1 | f16 | f16 | causalbr | no_quantization |             36 |           33 | 1.0909091 |

LNL:
| mb | Q Heads | KV Heads |   D |    K |    Q | Kdt | Vdt | mask     | quant           | base time (us) | PR time (us) |    speedup |
|----|---------|----------|-----|------|------|-----|-----|----------|-----------------|----------------|--------------|------------|
|  1 |      32 |        8 | 128 |   64 |   64 | f16 | f16 | causaltl | no_quantization |             53 |           52 |  1.0192308 |
|  1 |      32 |        8 | 128 |   65 |    1 | f16 | f16 | causaltl | no_quantization |             31 |           30 |  1.0333333 |
|  1 |      32 |        8 | 128 |  128 |  128 | f16 | f16 | causaltl | no_quantization |             77 |           74 |  1.0405405 |
|  1 |      32 |        8 | 128 |  129 |    1 | f16 | f16 | causaltl | no_quantization |             36 |           35 |  1.0285714 |
|  1 |      32 |        8 | 128 |  384 |  384 | f16 | f16 | causaltl | no_quantization |            308 |          216 |  1.4259259 |
|  1 |      32 |        8 | 128 |  385 |    1 | f16 | f16 | causaltl | no_quantization |             55 |           52 |  1.0576923 |
|  1 |      32 |        8 | 128 |  512 |  512 | f16 | f16 | causaltl | no_quantization |            432 |          296 |  1.4594595 |
|  1 |      32 |        8 | 128 |  513 |    1 | f16 | f16 | causaltl | no_quantization |             65 |           59 |  1.1016949 |
|  1 |      32 |        8 | 128 | 1024 | 1024 | f16 | f16 | causaltl | no_quantization |           1470 |          866 |  1.6974596 |
|  1 |      32 |        8 | 128 | 1025 |    1 | f16 | f16 | causaltl | no_quantization |            102 |          106 | 0.96226415 |
|  1 |      32 |        8 | 128 | 2048 | 2048 | f16 | f16 | causaltl | no_quantization |           5426 |         2796 |  1.9406295 |
|  1 |      32 |        8 | 128 | 2049 |    1 | f16 | f16 | causaltl | no_quantization |            170 |          166 |  1.0240964 |
|  1 |      32 |        8 | 128 |   64 |   64 | f16 | f16 | causalbr | no_quantization |             53 |           53 |          1 |
|  1 |      32 |        8 | 128 |   65 |    1 | f16 | f16 | causalbr | no_quantization |             31 |           31 |          1 |
|  1 |      32 |        8 | 128 |  128 |  128 | f16 | f16 | causalbr | no_quantization |             73 |           72 |  1.0138889 |
|  1 |      32 |        8 | 128 |  129 |    1 | f16 | f16 | causalbr | no_quantization |             34 |           35 | 0.97142857 |
|  1 |      32 |        8 | 128 |  384 |  384 | f16 | f16 | causalbr | no_quantization |            296 |          217 |  1.3640553 |
|  1 |      32 |        8 | 128 |  385 |    1 | f16 | f16 | causalbr | no_quantization |             52 |           58 | 0.89655172 |
|  1 |      32 |        8 | 128 |  512 |  512 | f16 | f16 | causalbr | no_quantization |            438 |          301 |  1.4551495 |
|  1 |      32 |        8 | 128 |  513 |    1 | f16 | f16 | causalbr | no_quantization |             69 |           64 |   1.078125 |
|  1 |      32 |        8 | 128 | 1024 | 1024 | f16 | f16 | causalbr | no_quantization |           1455 |          848 |  1.7158019 |
|  1 |      32 |        8 | 128 | 1025 |    1 | f16 | f16 | causalbr | no_quantization |            100 |           97 |  1.0309278 |
|  1 |      32 |        8 | 128 | 2048 | 2048 | f16 | f16 | causalbr | no_quantization |           5426 |         2787 |  1.9468963 |
|  1 |      32 |        8 | 128 | 2049 |    1 | f16 | f16 | causalbr | no_quantization |            168 |          165 |  1.0181818 |

A770:
| mb | Q Heads | KV Heads |   D |    K |    Q | Kdt | Vdt | mask     | quant           | base time (us) | PR time (us) |            |
|----|---------|----------|-----|------|------|-----|-----|----------|-----------------|----------------|--------------|------------|
|  1 |      32 |        8 | 128 |   64 |   64 | f16 | f16 | causaltl | no_quantization |             12 |           12 |          1 |
|  1 |      32 |        8 | 128 |   65 |    1 | f16 | f16 | causaltl | no_quantization |             14 |           15 | 0.93333333 |
|  1 |      32 |        8 | 128 |  128 |  128 | f16 | f16 | causaltl | no_quantization |             17 |           18 | 0.94444444 |
|  1 |      32 |        8 | 128 |  129 |    1 | f16 | f16 | causaltl | no_quantization |             21 |           21 |          1 |
|  1 |      32 |        8 | 128 |  384 |  384 | f16 | f16 | causaltl | no_quantization |            117 |           95 |  1.2315789 |
|  1 |      32 |        8 | 128 |  385 |    1 | f16 | f16 | causaltl | no_quantization |             32 |           32 |          1 |
|  1 |      32 |        8 | 128 |  512 |  512 | f16 | f16 | causaltl | no_quantization |            198 |          146 |  1.3561644 |
|  1 |      32 |        8 | 128 |  513 |    1 | f16 | f16 | causaltl | no_quantization |             40 |           40 |          1 |
|  1 |      32 |        8 | 128 | 1024 | 1024 | f16 | f16 | causaltl | no_quantization |           1070 |          494 |  2.1659919 |
|  1 |      32 |        8 | 128 | 1025 |    1 | f16 | f16 | causaltl | no_quantization |             72 |           72 |          1 |
|  1 |      32 |        8 | 128 | 2048 | 2048 | f16 | f16 | causaltl | no_quantization |           3524 |         1703 |  2.0692895 |
|  1 |      32 |        8 | 128 | 2049 |    1 | f16 | f16 | causaltl | no_quantization |            136 |          136 |          1 |
|  1 |      32 |        8 | 128 |   64 |   64 | f16 | f16 | causalbr | no_quantization |             12 |           12 |          1 |
|  1 |      32 |        8 | 128 |   65 |    1 | f16 | f16 | causalbr | no_quantization |             14 |           15 | 0.93333333 |
|  1 |      32 |        8 | 128 |  128 |  128 | f16 | f16 | causalbr | no_quantization |             17 |           18 | 0.94444444 |
|  1 |      32 |        8 | 128 |  129 |    1 | f16 | f16 | causalbr | no_quantization |             21 |           21 |          1 |
|  1 |      32 |        8 | 128 |  384 |  384 | f16 | f16 | causalbr | no_quantization |            117 |           94 |  1.2446809 |
|  1 |      32 |        8 | 128 |  385 |    1 | f16 | f16 | causalbr | no_quantization |             32 |           32 |          1 |
|  1 |      32 |        8 | 128 |  512 |  512 | f16 | f16 | causalbr | no_quantization |            198 |          145 |  1.3655172 |
|  1 |      32 |        8 | 128 |  513 |    1 | f16 | f16 | causalbr | no_quantization |             40 |           40 |          1 |
|  1 |      32 |        8 | 128 | 1024 | 1024 | f16 | f16 | causalbr | no_quantization |           1070 |          494 |  2.1659919 |
|  1 |      32 |        8 | 128 | 1025 |    1 | f16 | f16 | causalbr | no_quantization |             72 |           72 |          1 |
|  1 |      32 |        8 | 128 | 2048 | 2048 | f16 | f16 | causalbr | no_quantization |           3521 |         1701 |  2.0699588 |
|  1 |      32 |        8 | 128 | 2049 |    1 | f16 | f16 | causalbr | no_quantization |            136 |          135 |  1.0074074 |